Issue#120 fix memory leak in busmap
